MIPS: Make sure _start symbol is defined as type FUNC
Explicitly set _start symbol type as a function.
Additionaly, make sure _start has an address different than __start
by adding one nop in between.

This patch fixes an issue reported for LLDB where two symbols
_start & __start with different types have the same address:

...
00000510    52 FUNC    GLOBAL DEFAULT    9 __start
00000510     0 OBJECT  GLOBAL DEFAULT    9 _start
...

This is causing problem for LLDB in finding information of shared libraries.
LLDB ends up storing symbol of type as OBJECT at 00000510 (because of the
sequence it appears in elf) and can not place a breakpoint at this address
since it finds a data symbol there (and not a function symbol).

Change-Id: Ida43ee8e5960cc771a4f3b9b5d7f8010e90c4163
Reported-by: Bhushan Attarde <bhushan.attarde@imgtec.com>
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>